revert: move wallet endpoints back to original locations
Reverts the endpoint moves from PR #131 to save for a different PR:
- POST /gateway/wallets/create (stays in gateway.py)
- POST /gateway/wallets/show-private-key (stays in gateway.py)
- POST /gateway/wallets/send (stays in gateway.py)
- GET /accounts/gateway/wallets (back to accounts.py)
- POST /accounts/gateway/add-wallet (back to accounts.py)
- DELETE /accounts/gateway/{chain}/{address} (back to accounts.py)
- POST /accounts/gateway/wallet/set-default (new, in accounts.py)
