+++ routers/accounts.py
@@ -1,20 +1,11 @@
-from typing import Dict, List, Optional
-from datetime import datetime
+from typing import Dict, List
 
-from fastapi import APIRouter, HTTPException, Depends, Query
+from fastapi import APIRouter, Depends, HTTPException
 from starlette import status
 
-from services.accounts_service import AccountsService
 from deps import get_accounts_service
-from models import (
-    PaginatedResponse,
-    GatewayWalletCredential,
-    GatewayWalletInfo,
-    CreateWalletRequest,
-    ShowPrivateKeyRequest,
-    SendTransactionRequest,
-    SetDefaultWalletRequest,
-)
+from models import GatewayWalletCredential, SetDefaultWalletRequest
+from services.accounts_service import AccountsService
 
 router = APIRouter(tags=["Accounts"], prefix="/accounts")
 
@@ -23,7 +14,7 @@ router = APIRouter(tags=["Accounts"], prefix="/accounts")
 async def list_accounts(accounts_service: AccountsService = Depends(get_accounts_service)):
     """
     Get a list of all account names in the system.
-    
+
     Returns:
         List of account names
     """
@@ -59,13 +50,13 @@ async def list_account_credentials(account_name: str,
 async def add_account(account_name: str, accounts_service: AccountsService = Depends(get_accounts_service)):
     """
     Create a new account with default configuration files.
-    
+
     Args:
         account_name: Name of the new account to create
-        
+
     Returns:
         Success message when account is created
-        
+
     Raises:
         HTTPException: 400 if account already exists
     """
@@ -80,13 +71,13 @@ async def add_account(account_name: str, accounts_service: AccountsService = Dep
 async def delete_account(account_name: str, accounts_service: AccountsService = Depends(get_accounts_service)):
     """
     Delete an account and all its associated credentials.
-    
+
     Args:
         account_name: Name of the account to delete
-        
+
     Returns:
         Success message when account is deleted
-        
+
     Raises:
         HTTPException: 400 if trying to delete master account, 404 if account not found
     """
@@ -103,14 +94,14 @@ async def delete_account(account_name: str, accounts_service: AccountsService = 
 async def delete_credential(account_name: str, connector_name: str, accounts_service: AccountsService = Depends(get_accounts_service)):
     """
     Delete a specific connector credential for an account.
-    
+
     Args:
         account_name: Name of the account
         connector_name: Name of the connector to delete credentials for
-        
+
     Returns:
         Success message when credential is deleted
-        
+
     Raises:
         HTTPException: 404 if credential not found
     """
@@ -170,7 +161,7 @@ async def list_gateway_wallets(accounts_service: AccountsService = Depends(get_a
         raise HTTPException(status_code=500, detail=str(e))
 
 
-@router.post("/gateway/wallet/add", status_code=status.HTTP_201_CREATED)
+@router.post("/gateway/add-wallet", status_code=status.HTTP_201_CREATED)
 async def add_gateway_wallet(
     wallet_credential: GatewayWalletCredential,
     accounts_service: AccountsService = Depends(get_accounts_service)
@@ -252,7 +243,7 @@ async def set_default_gateway_wallet(
         raise HTTPException(status_code=500, detail=f"Error setting default wallet: {str(e)}")
 
 
-@router.delete("/gateway/wallet/{chain}/{address}")
+@router.delete("/gateway/{chain}/{address}")
 async def remove_gateway_wallet(
     chain: str,
     address: str,
@@ -278,142 +269,3 @@ async def remove_gateway_wallet(
         raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
-
-@router.post("/gateway/wallet/create", status_code=status.HTTP_201_CREATED)
-async def create_gateway_wallet(
-    request: CreateWalletRequest,
-    accounts_service: AccountsService = Depends(get_accounts_service)
-) -> Dict:
-    """
-    Create a new wallet in Gateway.
-
-    Args:
-        request: Contains chain and set_default flag
-
-    Returns:
-        Dict with address and chain of the created wallet.
-
-    Example: POST /accounts/gateway/wallet/create
-    {
-        "chain": "solana",
-        "set_default": true
-    }
-    """
-    try:
-        if not await accounts_service.gateway_client.ping():
-            raise HTTPException(status_code=503, detail="Gateway service is not available")
-
-        result = await accounts_service.gateway_client.create_wallet(
-            chain=request.chain,
-            set_default=request.set_default
-        )
-
-        if result is None:
-            raise HTTPException(status_code=502, detail="Failed to create wallet: Gateway returned no response")
-
-        if "error" in result:
-            raise HTTPException(status_code=400, detail=f"Failed to create wallet: {result.get('error')}")
-
-        return result
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Error creating wallet: {str(e)}")
-
-
-@router.post("/gateway/wallet/show-private-key")
-async def show_gateway_wallet_private_key(
-    request: ShowPrivateKeyRequest,
-    accounts_service: AccountsService = Depends(get_accounts_service)
-) -> Dict:
-    """
-    Show private key for a wallet.
-
-    WARNING: This endpoint exposes sensitive information. Use with caution.
-
-    Args:
-        request: Contains chain, address, and passphrase
-
-    Returns:
-        Dict with privateKey field.
-
-    Example: POST /accounts/gateway/wallet/show-private-key
-    {
-        "chain": "solana",
-        "address": "<wallet-address>",
-        "passphrase": "<gateway-passphrase>"
-    }
-    """
-    try:
-        if not await accounts_service.gateway_client.ping():
-            raise HTTPException(status_code=503, detail="Gateway service is not available")
-
-        result = await accounts_service.gateway_client.show_private_key(
-            chain=request.chain,
-            address=request.address,
-            passphrase=request.passphrase
-        )
-
-        if result is None:
-            raise HTTPException(status_code=502, detail="Failed to retrieve private key: Gateway returned no response")
-
-        if "error" in result:
-            raise HTTPException(status_code=400, detail=f"Failed to retrieve private key: {result.get('error')}")
-
-        return result
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Error retrieving private key: {str(e)}")
-
-
-@router.post("/gateway/wallet/send")
-async def send_gateway_wallet_transaction(
-    request: SendTransactionRequest,
-    accounts_service: AccountsService = Depends(get_accounts_service)
-) -> Dict:
-    """
-    Send a native token transaction.
-
-    Args:
-        request: Contains chain, network, sender address, recipient address, and amount
-
-    Returns:
-        Dict with transaction signature/hash.
-
-    Example: POST /accounts/gateway/wallet/send
-    {
-        "chain": "solana",
-        "network": "mainnet-beta",
-        "address": "<sender-address>",
-        "to_address": "<recipient-address>",
-        "amount": "0.001"
-    }
-    """
-    try:
-        if not await accounts_service.gateway_client.ping():
-            raise HTTPException(status_code=503, detail="Gateway service is not available")
-
-        result = await accounts_service.gateway_client.send_transaction(
-            chain=request.chain,
-            network=request.network,
-            address=request.address,
-            to_address=request.to_address,
-            amount=request.amount
-        )
-
-        if result is None:
-            raise HTTPException(status_code=502, detail="Failed to send transaction: Gateway returned no response")
-
-        if "error" in result:
-            raise HTTPException(status_code=400, detail=f"Failed to send transaction: {result.get('error')}")
-
-        return result
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Error sending transaction: {str(e)}")
