+++ routers/gateway.py
@@ -1,16 +1,20 @@
-from fastapi import APIRouter, HTTPException, Depends, Query
-from typing import Optional, Dict, List
 import re
+from typing import Dict, List, Optional
 
+from fastapi import APIRouter, Depends, HTTPException, Query
+
+from deps import get_accounts_service, get_gateway_service
 from models import (
-    GatewayConfig,
-    GatewayStatus,
     AddPoolRequest,
     AddTokenRequest,
+    CreateWalletRequest,
+    GatewayConfig,
+    GatewayStatus,
+    SendTransactionRequest,
+    ShowPrivateKeyRequest,
 )
-from services.gateway_service import GatewayService
 from services.accounts_service import AccountsService
-from deps import get_gateway_service, get_accounts_service
+from services.gateway_service import GatewayService
 
 router = APIRouter(tags=["Gateway"], prefix="/gateway")
 
@@ -668,3 +672,146 @@ async def delete_network_token(
         raise
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error deleting token: {str(e)}")
+
+
+# ============================================
+# Wallet Management
+# ============================================
+
+@router.post("/wallets/create")
+async def create_gateway_wallet(
+    request: CreateWalletRequest,
+    accounts_service: AccountsService = Depends(get_accounts_service)
+) -> Dict:
+    """
+    Create a new wallet in Gateway.
+
+    Args:
+        request: Contains chain and set_default flag
+
+    Returns:
+        Dict with address and chain of the created wallet.
+
+    Example: POST /gateway/wallets/create
+    {
+        "chain": "solana",
+        "set_default": true
+    }
+    """
+    try:
+        if not await accounts_service.gateway_client.ping():
+            raise HTTPException(status_code=503, detail="Gateway service is not available")
+
+        result = await accounts_service.gateway_client.create_wallet(
+            chain=request.chain,
+            set_default=request.set_default
+        )
+
+        if result is None:
+            raise HTTPException(status_code=502, detail="Failed to create wallet: Gateway returned no response")
+
+        if "error" in result:
+            raise HTTPException(status_code=400, detail=f"Failed to create wallet: {result.get('error')}")
+
+        return result
+
+    except HTTPException:
+        raise
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Error creating wallet: {str(e)}")
+
+
+@router.post("/wallets/show-private-key")
+async def show_gateway_wallet_private_key(
+    request: ShowPrivateKeyRequest,
+    accounts_service: AccountsService = Depends(get_accounts_service)
+) -> Dict:
+    """
+    Show private key for a wallet.
+
+    WARNING: This endpoint exposes sensitive information. Use with caution.
+
+    Args:
+        request: Contains chain, address, and passphrase
+
+    Returns:
+        Dict with privateKey field.
+
+    Example: POST /gateway/wallets/show-private-key
+    {
+        "chain": "solana",
+        "address": "<wallet-address>",
+        "passphrase": "<gateway-passphrase>"
+    }
+    """
+    try:
+        if not await accounts_service.gateway_client.ping():
+            raise HTTPException(status_code=503, detail="Gateway service is not available")
+
+        result = await accounts_service.gateway_client.show_private_key(
+            chain=request.chain,
+            address=request.address,
+            passphrase=request.passphrase
+        )
+
+        if result is None:
+            raise HTTPException(status_code=502, detail="Failed to retrieve private key: Gateway returned no response")
+
+        if "error" in result:
+            raise HTTPException(status_code=400, detail=f"Failed to retrieve private key: {result.get('error')}")
+
+        return result
+
+    except HTTPException:
+        raise
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Error retrieving private key: {str(e)}")
+
+
+@router.post("/wallets/send")
+async def send_gateway_wallet_transaction(
+    request: SendTransactionRequest,
+    accounts_service: AccountsService = Depends(get_accounts_service)
+) -> Dict:
+    """
+    Send a native token transaction.
+
+    Args:
+        request: Contains chain, network, sender address, recipient address, and amount
+
+    Returns:
+        Dict with transaction signature/hash.
+
+    Example: POST /gateway/wallets/send
+    {
+        "chain": "solana",
+        "network": "mainnet-beta",
+        "address": "<sender-address>",
+        "to_address": "<recipient-address>",
+        "amount": "0.001"
+    }
+    """
+    try:
+        if not await accounts_service.gateway_client.ping():
+            raise HTTPException(status_code=503, detail="Gateway service is not available")
+
+        result = await accounts_service.gateway_client.send_transaction(
+            chain=request.chain,
+            network=request.network,
+            address=request.address,
+            to_address=request.to_address,
+            amount=request.amount
+        )
+
+        if result is None:
+            raise HTTPException(status_code=502, detail="Failed to send transaction: Gateway returned no response")
+
+        if "error" in result:
+            raise HTTPException(status_code=400, detail=f"Failed to send transaction: {result.get('error')}")
+
+        return result
+
+    except HTTPException:
+        raise
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Error sending transaction: {str(e)}")
